wgpu: Move descriptors out into its own file
