+++ exporter/src/main.rs
@@ -6,8 +6,9 @@ use rayon::prelude::*;
 use ruffle_core::tag_utils::SwfMovie;
 use ruffle_core::PlayerBuilder;
 use ruffle_render_wgpu::clap::{GraphicsBackend, PowerPreference};
+use ruffle_render_wgpu::descriptors::Descriptors;
 use ruffle_render_wgpu::target::TextureTarget;
-use ruffle_render_wgpu::{wgpu, Descriptors, WgpuRenderBackend};
+use ruffle_render_wgpu::{wgpu, WgpuRenderBackend};
 use std::fs::create_dir_all;
 use std::panic::catch_unwind;
 use std::path::{Path, PathBuf};
